WIP: move choices
