+++ src/hope/api/endpoints/lookups/base.py
@@ -11,14 +11,16 @@ from hope.api.endpoints.serializers import (
     FinancialInstitutionListSerializer,
 )
 from hope.api.filters import CountryFilter, FinancialInstitutionFilter
-from hope.apps.household.const import RESIDENCE_STATUS_CHOICE
-from hope.models import (
+from hope.apps.household.const import (
     IDENTIFICATION_TYPE_CHOICE,
     MARITAL_STATUS_CHOICE,
     OBSERVED_DISABILITY_CHOICE,
     RELATIONSHIP_CHOICE,
+    RESIDENCE_STATUS_CHOICE,
     ROLE_CHOICE,
     SEX_CHOICE,
+)
+from hope.models import (
     Country,
     FinancialInstitution,
     Program,
