+++ src/hope/models/household.py
@@ -41,251 +41,20 @@ from hope.models.utils import (
 )
 
 if TYPE_CHECKING:
-    from django.utils.functional import _StrPromise
-
     from hope.contrib.aurora.models import Record
     from hope.models.area import Area
     from hope.models.business_area import BusinessArea
     from hope.models.individual import Individual
 
-# TODO: remove later
-# please use hope.apps.household.const
-BLANK = ""
-# INDIVIDUALS
-MALE = "MALE"
-FEMALE = "FEMALE"
-OTHER = "OTHER"
-NOT_COLLECTED = "NOT_COLLECTED"
-NOT_ANSWERED = "NOT_ANSWERED"
-
-SEX_CHOICE: tuple[tuple[str, _StrPromise], ...] = (
-    (MALE, _("Male")),
-    (FEMALE, _("Female")),
-    (OTHER, _("Other")),
-    (NOT_COLLECTED, _("Not collected")),
-    (NOT_ANSWERED, _("Not answered")),
-)
-
-SINGLE = "SINGLE"
-MARRIED = "MARRIED"
-WIDOWED = "WIDOWED"
-DIVORCED = "DIVORCED"
-SEPARATED = "SEPARATED"
-MARITAL_STATUS_CHOICE: tuple[tuple[str, _StrPromise], ...] = (
-    (BLANK, _("None")),
-    (DIVORCED, _("Divorced")),
-    (MARRIED, _("Married")),
-    (SEPARATED, _("Separated")),
-    (SINGLE, _("Single")),
-    (WIDOWED, _("Widowed")),
-)
-
-NONE = "NONE"
-SEEING = "SEEING"
-HEARING = "HEARING"
-WALKING = "WALKING"
-MEMORY = "MEMORY"
-SELF_CARE = "SELF_CARE"
-COMMUNICATING = "COMMUNICATING"
-OBSERVED_DISABILITY_CHOICE: tuple[tuple[str, _StrPromise], ...] = (
-    (NONE, _("None")),
-    (SEEING, _("Difficulty seeing (even if wearing glasses)")),
-    (HEARING, _("Difficulty hearing (even if using a hearing aid)")),
-    (WALKING, _("Difficulty walking or climbing steps")),
-    (MEMORY, _("Difficulty remembering or concentrating")),
-    (SELF_CARE, _("Difficulty with self care (washing, dressing)")),
-    (
-        COMMUNICATING,
-        _("Difficulty communicating (e.g understanding or being understood)"),
-    ),
-)
-NON_BENEFICIARY = "NON_BENEFICIARY"  # delegate
-HEAD = "HEAD"
-SON_DAUGHTER = "SON_DAUGHTER"
-WIFE_HUSBAND = "WIFE_HUSBAND"
-BROTHER_SISTER = "BROTHER_SISTER"
-MOTHER_FATHER = "MOTHER_FATHER"
-AUNT_UNCLE = "AUNT_UNCLE"
-GRANDMOTHER_GRANDFATHER = "GRANDMOTHER_GRANDFATHER"
-MOTHERINLAW_FATHERINLAW = "MOTHERINLAW_FATHERINLAW"
-DAUGHTERINLAW_SONINLAW = "DAUGHTERINLAW_SONINLAW"
-SISTERINLAW_BROTHERINLAW = "SISTERINLAW_BROTHERINLAW"
-GRANDDAUGHTER_GRANDSON = "GRANDDAUGHER_GRANDSON"  # key is wrong, but it is used in kobo and aurora
-NEPHEW_NIECE = "NEPHEW_NIECE"
-COUSIN = "COUSIN"
-FOSTER_CHILD = "FOSTER_CHILD"
-RELATIONSHIP_UNKNOWN = "UNKNOWN"
-RELATIONSHIP_OTHER = "OTHER"
-FREE_UNION = "FREE_UNION"
-
-RELATIONSHIP_CHOICE: tuple[tuple[str, str], ...] = (
-    (RELATIONSHIP_UNKNOWN, "Unknown"),
-    (AUNT_UNCLE, "Aunt / Uncle"),
-    (BROTHER_SISTER, "Brother / Sister"),
-    (COUSIN, "Cousin"),
-    (DAUGHTERINLAW_SONINLAW, "Daughter-in-law / Son-in-law"),
-    (GRANDDAUGHTER_GRANDSON, "Granddaughter / Grandson"),
-    (GRANDMOTHER_GRANDFATHER, "Grandmother / Grandfather"),
-    (HEAD, "Head of household (self)"),
-    (MOTHER_FATHER, "Mother / Father"),
-    (MOTHERINLAW_FATHERINLAW, "Mother-in-law / Father-in-law"),
-    (NEPHEW_NIECE, "Nephew / Niece"),
-    (
-        NON_BENEFICIARY,
-        "Not a Family Member. Can only act as a recipient.",
-    ),
-    (RELATIONSHIP_OTHER, "Other"),
-    (SISTERINLAW_BROTHERINLAW, "Sister-in-law / Brother-in-law"),
-    (SON_DAUGHTER, "Son / Daughter"),
-    (WIFE_HUSBAND, "Wife / Husband"),
-    (FOSTER_CHILD, "Foster child"),
-    (FREE_UNION, "Free union"),
-)
-YES = "1"
-NO = "0"
-YES_NO_CHOICE: tuple[tuple[str, _StrPromise], ...] = (
-    (BLANK, _("None")),
-    (YES, _("Yes")),
-    (NO, _("No")),
-)
-
-NOT_PROVIDED = "NOT_PROVIDED"
-WORK_STATUS_CHOICE: tuple[tuple[str, _StrPromise], ...] = (
-    (YES, _("Yes")),
-    (NO, _("No")),
-    (NOT_PROVIDED, _("Not provided")),
-)
-ROLE_PRIMARY = "PRIMARY"
-ROLE_ALTERNATE = "ALTERNATE"
-ROLE_CHOICE: tuple[tuple[str, str], ...] = (
-    (ROLE_ALTERNATE, "Alternate collector"),
-    (ROLE_PRIMARY, "Primary collector"),
-)
-IDENTIFICATION_TYPE_BIRTH_CERTIFICATE = "BIRTH_CERTIFICATE"
-IDENTIFICATION_TYPE_DRIVERS_LICENSE = "DRIVERS_LICENSE"
-IDENTIFICATION_TYPE_NATIONAL_ID = "NATIONAL_ID"
-IDENTIFICATION_TYPE_NATIONAL_PASSPORT = "NATIONAL_PASSPORT"
-IDENTIFICATION_TYPE_ELECTORAL_CARD = "ELECTORAL_CARD"
-IDENTIFICATION_TYPE_TAX_ID = "TAX_ID"
-IDENTIFICATION_TYPE_RESIDENCE_PERMIT_NO = "RESIDENCE_PERMIT_NO"
-IDENTIFICATION_TYPE_BANK_STATEMENT = "BANK_STATEMENT"
-IDENTIFICATION_TYPE_DISABILITY_CERTIFICATE = "DISABILITY_CERTIFICATE"
-IDENTIFICATION_TYPE_FOSTER_CHILD = "FOSTER_CHILD"
-IDENTIFICATION_TYPE_OTHER = "OTHER"
-IDENTIFICATION_TYPE_CHOICE = (
-    (IDENTIFICATION_TYPE_BIRTH_CERTIFICATE, _("Birth Certificate")),
-    (IDENTIFICATION_TYPE_DRIVERS_LICENSE, _("Driver's License")),
-    (IDENTIFICATION_TYPE_ELECTORAL_CARD, _("Electoral Card")),
-    (IDENTIFICATION_TYPE_NATIONAL_ID, _("National ID")),
-    (IDENTIFICATION_TYPE_NATIONAL_PASSPORT, _("National Passport")),
-    (IDENTIFICATION_TYPE_TAX_ID, _("Tax Number Identification")),
-    (IDENTIFICATION_TYPE_RESIDENCE_PERMIT_NO, _("Foreigner's Residence Permit")),
-    (IDENTIFICATION_TYPE_BANK_STATEMENT, _("Bank Statement")),
-    (IDENTIFICATION_TYPE_DISABILITY_CERTIFICATE, _("Disability Certificate")),
-    (IDENTIFICATION_TYPE_FOSTER_CHILD, _("Foster Child")),
-    (IDENTIFICATION_TYPE_OTHER, _("Other")),
-)
-IDENTIFICATION_TYPE_DICT = {
-    IDENTIFICATION_TYPE_BIRTH_CERTIFICATE: "Birth Certificate",
-    IDENTIFICATION_TYPE_DRIVERS_LICENSE: "Driver's License",
-    IDENTIFICATION_TYPE_ELECTORAL_CARD: "Electoral Card",
-    IDENTIFICATION_TYPE_NATIONAL_ID: "National ID",
-    IDENTIFICATION_TYPE_NATIONAL_PASSPORT: "National Passport",
-    IDENTIFICATION_TYPE_TAX_ID: "Tax Number Identification",
-    IDENTIFICATION_TYPE_RESIDENCE_PERMIT_NO: "Foreigner's Residence Permit",
-    IDENTIFICATION_TYPE_OTHER: "Other",
-}
-UNHCR = "UNHCR"
-WFP = "WFP"
-AGENCY_TYPE_CHOICES = (
-    (UNHCR, _("UNHCR")),
-    (WFP, _("WFP")),
-)
-STATUS_ACTIVE = "ACTIVE"
-STATUS_INACTIVE = "INACTIVE"
-STATUS_WITHDRAWN = "WITHDRAWN"
-STATUS_DUPLICATE = "DUPLICATE"
-INDIVIDUAL_STATUS_CHOICES = (
-    (STATUS_ACTIVE, "Active"),
-    (STATUS_DUPLICATE, "Duplicate"),
-    (STATUS_WITHDRAWN, "Withdrawn"),
-)
-INDIVIDUAL_HOUSEHOLD_STATUS = ((STATUS_ACTIVE, "Active"), (STATUS_INACTIVE, "Inactive"))
-UNIQUE = "UNIQUE"
-DUPLICATE = "DUPLICATE"
-NEEDS_ADJUDICATION = "NEEDS_ADJUDICATION"
-NOT_PROCESSED = "NOT_PROCESSED"
-DEDUPLICATION_POSTPONE = "POSTPONE"
-DEDUPLICATION_GOLDEN_RECORD_STATUS_CHOICE: tuple[tuple[str, str], ...] = (
-    (DUPLICATE, "Duplicate"),
-    (NEEDS_ADJUDICATION, "Needs Adjudication"),
-    (NOT_PROCESSED, "Not Processed"),
-    (DEDUPLICATION_POSTPONE, "Postpone"),
-    (UNIQUE, "Unique"),
-)
-SIMILAR_IN_BATCH = "SIMILAR_IN_BATCH"
-DUPLICATE_IN_BATCH = "DUPLICATE_IN_BATCH"
-UNIQUE_IN_BATCH = "UNIQUE_IN_BATCH"
-NOT_PROCESSED = "NOT_PROCESSED"
-DEDUPLICATION_BATCH_STATUS_CHOICE: tuple[tuple[str, str], ...] = (
-    (DUPLICATE_IN_BATCH, "Duplicate in batch"),
-    (NOT_PROCESSED, "Not Processed"),
-    (SIMILAR_IN_BATCH, "Similar in batch"),
-    (UNIQUE_IN_BATCH, "Unique in batch"),
-)
-SOME_DIFFICULTY = "SOME_DIFFICULTY"
-LOT_DIFFICULTY = "LOT_DIFFICULTY"
-CANNOT_DO = "CANNOT_DO"
-SEVERITY_OF_DISABILITY_CHOICES: tuple[tuple[str, str | _StrPromise], ...] = (
-    (BLANK, _("None")),
-    (LOT_DIFFICULTY, "A lot of difficulty"),
-    (CANNOT_DO, "Cannot do at all"),
-    (SOME_DIFFICULTY, "Some difficulty"),
-)
-UNICEF = "UNICEF"
-PARTNER = "PARTNER"
-ORG_ENUMERATOR_CHOICES = (
-    (BLANK, _("None")),
-    (PARTNER, "Partner"),
-    (UNICEF, "UNICEF"),
-)
-HUMANITARIAN_PARTNER = "HUMANITARIAN_PARTNER"
-PRIVATE_PARTNER = "PRIVATE_PARTNER"
-GOVERNMENT_PARTNER = "GOVERNMENT_PARTNER"
-DATA_SHARING_CHOICES = (
-    (BLANK, _("None")),
-    (GOVERNMENT_PARTNER, "Government partners"),
-    (HUMANITARIAN_PARTNER, "Humanitarian partners"),
-    (PRIVATE_PARTNER, "Private partners"),
-    (UNICEF, "UNICEF"),
-)
-HH_REGISTRATION = "HH_REGISTRATION"
-COMMUNITY = "COMMUNITY"
-REGISTRATION_METHOD_CHOICES = (
-    (BLANK, _("None")),
-    (COMMUNITY, "Community-level Registration"),
-    (HH_REGISTRATION, "Household Registration"),
-)
-
-DISABLED = "disabled"
-NOT_DISABLED = "not disabled"
-DISABILITY_CHOICES: tuple[tuple[str, str], ...] = (
-    (
-        DISABLED,
-        "disabled",
-    ),
-    (
-        NOT_DISABLED,
-        "not disabled",
-    ),
-)
-SANCTION_LIST_POSSIBLE_MATCH = "SANCTION_LIST_POSSIBLE_MATCH"
-SANCTION_LIST_CONFIRMED_MATCH = "SANCTION_LIST_CONFIRMED_MATCH"
-INDIVIDUAL_FLAGS_CHOICES = (
-    (DUPLICATE, "Duplicate"),
-    (NEEDS_ADJUDICATION, "Needs adjudication"),
-    (SANCTION_LIST_CONFIRMED_MATCH, "Sanction list match"),
-    (SANCTION_LIST_POSSIBLE_MATCH, "Sanction list possible match"),
+from hope.apps.household.const import (
+    BLANK,
+    DATA_SHARING_CHOICES,
+    ORG_ENUMERATOR_CHOICES,
+    REGISTRATION_METHOD_CHOICES,
+    ROLE_ALTERNATE,
+    ROLE_PRIMARY,
+    STATUS_ACTIVE,
+    STATUS_INACTIVE,
 )
 
 logger = logging.getLogger(__name__)
