+++ tests/unit/apps/core/test_attributes_get_queries.py
@@ -40,7 +40,7 @@ from hope.apps.core.attributes_qet_queries import (
     registration_data_import_query,
 )
 from hope.apps.core.countries import Countries
-from hope.models import UNHCR, WFP
+from hope.apps.household.const import UNHCR, WFP
 
 TODAY = dt.date.today()
